[ADD] adding sigmoid

+++ include/layers/linear/linear.hpp
@@ -11,19 +11,19 @@ public:
 
     int get_input_dim(void);
     int get_output_dim(void);
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> get_w(void);
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> get_b(void);
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> get_w(void);
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> get_b(void);
     void compute_logits(Eigen::MatrixXf input); // change to be reference and not copy
 
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> logits;
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> logits;
 
 private:
     const int _input_dim;
     const int _output_dim;
     const int _n_rows;
 
-    Eigen::Matrix<float, 1, Eigen::Dynamic> _b;
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> _w;
+    Eigen::Matrix<double, 1, Eigen::Dynamic> _b;
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> _w;
     // Eigen::MatrixXf _w;
     // Eigen::MatrixXf _b;  
 };
